moved old class objects to raw_dataset.py and renamed meth_dataset to sigset.py; dropped unused stuff from __init__ for each folder
